requirements.txt
streamlit>=1.40,<2.0
matplotlib>=3.8,<4.0
pandas>=2.2,<3.0
plotly>=5.24,<6.0
python-dotenv>=1.0,<2.0
